pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spike-utils"
version = "0.1.0"
description = "Utilities for analyzing spiking neural data."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "scipy",
    "matplotlib",
    "pandas",
    "tqdm",
    "cupy-cuda12x",
    "torch>=2.0",
]

[tool.setuptools.packages.find]
where = ["."]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
